commit 2

+++ utils/src/main/java/com/epam/utils/StringUtils.java
@@ -2,7 +2,7 @@ package com.epam.utils;
 
 public class StringUtils {
     public static boolean isPositiveNumber(String str) {
-        return Integer.parseInt(str) > 0;
+        return Float.parseFloat(str) > 0;
          //here magic will happen
     }
 }
